requirements.txt
dash
pandas
matplotlib
tensorflow-cpu
Pillow
plotly
gunicorn
kaleido